+++ packages/opds-web-client/CHANGELOG.md
@@ -2,9 +2,9 @@
 
 ### v0.3.2
 
-- Pass the redux store down the tree via context (this was not done previously)
+- Pass the redux store down the tree via context
 - Extracted PathForContext from the Root, which provides context via both legacy and new context APIs
-- Created a RouterContext component to be used by `circulation-patron-web` to provide router context.
+- Created a RouterContext component which can be used by any application to pass a `router` context down the tree. See `circulation-patron-web` as an example.
 
 ### v0.3.1